+++ src/search.cpp
@@ -2,9 +2,10 @@
 #include "search.hpp"
 #endif
 
-usize b_src_on_array(uint *arr, usize n, uint element, usize upper_bound) {
-    usize left = 0;
-    usize right = upper_bound;
+uintptr_t b_src_on_array(int64_t *arr, uintptr_t n, int64_t element,
+                         uintptr_t upper_bound) {
+    uintptr_t left = 0;
+    uintptr_t right = upper_bound;
 
     // Out of bounds check
     if (upper_bound >= n) {
